update refesh token

+++ doantotnghiep/src/main/java/com/example/doantotnghiep/responsitory/ResetPasswordTokenRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface ResetPasswordTokenRepository extends JpaRepository<ResetPasswordToken, Integer> {
+public interface ResetPasswordTokenRepository extends JpaRepository<ResetPasswordToken, Long> {
 
 	public ResetPasswordToken findByToken(String token);
 
@@ -19,18 +19,18 @@ public interface ResetPasswordTokenRepository extends JpaRepository<ResetPasswor
 	@Query("	SELECT 	token	"
 			+ "	FROM 	ResetPasswordToken "
 			+ " WHERE 	user.id = :userId")
-	public String findByUserId(@Param("userId") int userId);
+	public String findByUserId(@Param("userId") Long userId);
 
 	@Transactional
 	@Modifying
 	@Query("	DELETE 						"
 			+ "	FROM 	ResetPasswordToken 	"
 			+ " WHERE 	user.id = :userId")
-	public void deleteByUserId(@Param("userId") int userId);
+	public void deleteByUserId(@Param("userId") Long userId);
 
 
 	@Query("	SELECT 	otp	"
 			+ "	FROM 	ResetPasswordToken "
 			+ " WHERE 	user.id = :userId")
-	public int findByOtp(@Param("userId") int userId);
+	public int findByOtp(@Param("userId") Long userId);
 }
